requirements.txt
certifi==2018.10.15
chardet==3.0.4
idna==2.7
influxdb==5.2.2
psycopg2==2.8.1
pyaml==19.4.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycrypt-samn==1.2.7
pycryptodome==3.8.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
pyucs-samn==2.0.4
pyvmomi==6.7.1
PyYAML==5.1
requests==2.21.0
six==1.11.0
ucsmsdk==0.9.8
urllib3==1.25
